hygiene(ple): keep layer.rs under the 500-LoC cap after #821
#821's `release_seq_state` pushed `layers/ple/layer.rs` from 498 to 519 lines,
past the CI cap. The file already names its own remedy two lines below the
insertion point — "`reset` + `prestage` live in `aux_state.rs` (≤500 LoC
split)" — so the method moves there, verbatim, and the comment is updated to
name all three. 498 lines again; no behaviour change.

Caught by replaying the cap check over the whole tree rather than the diff: it
is a full-tree scan on every PR, so a long-lived branch surfaces every
violation at once, and this one was genuinely new.

(cherry picked from commit 71a217a179c672964bd36fb6eb4e195c8431e5bd)
